feature: Decouple prices evaluation to separate service
